requirements.txt
# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.0.196

# Computer Vision
opencv-python>=4.5.3.56
pillow>=8.3.0

# Data Processing
numpy>=1.22.2
pandas>=2.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Web Framework
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
aiofiles==0.7.0

# Dashboard
streamlit>=1.0.0

# Utilities
requests>=2.26.0
pyyaml>=6.0.1
tqdm>=4.62.0
jinja2==3.0.1
sqlalchemy==1.4.23
python-jose==3.3.0
python-dotenv==0.19.0
supervision>=0.16.0
databases[postgresql]==0.7.0
aiosqlite==0.17.0
bcrypt==3.2.0
python-socketio==5.4.0
websockets==10.0

# Added from the code block
flask==2.0.1
werkzeug==2.0.1
pytest==6.2.5
pytest-cov==2.12.1
black==21.7b0
flake8==3.9.2
mypy==0.910
isort==5.9.3
gunicorn==20.1.0
psycopg2-binary==2.9.9